Add files via upload
Project:
Business Objective:
A combined-cycle power plant comprises gas turbines, steam turbines, and heat recovery steam generators. In this type of plant, the electricity is generated by gas and steam turbines combined in one cycle. Then, it is transferred from one turbine to another. We have to model the energy generated as a function of exhaust vacuum and ambient variables and use that model to improve the plant's performance. 
Data Set Details: 
This is a project where the variable to be predicted is energy production
The data file contains 9568 observations with five variables collected from a combined cycle power plant over six years when the power plant was set to work with a full load. 

The variables, or features, are the following:
temperature, in degrees Celsius.
exhaust_vacuum, in cm Hg.
amb_pressure, in millibar. (Ambient pressure)
r_humidity, in percentage. (Relative humidity)
energy_production, in MW, net hourly electrical energy output.

Acceptance Criterion: Need to deploy the end results using Flask /Streamlit etc